+++ changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Prevent websocket reconnect loop when the authentication fails.
+* Prevent websocket reconnect when the connection closes without error.
 * Added gunicorn worker test to the CI pipeline.
 
 ## Released (2021-03-01)